simple and parallel B&B implementation

+++ Algorithm/heuristics.c
@@ -6,7 +6,7 @@
 #include "../DataStructure/DataStructure.h"
 #include "../utils.h"
 
-double assignmentCost(qa_global params, const int *assignment);
+int assignmentCost(qa_global params, const int *assignment);
 
 void swap(int *a, int *b);
 
@@ -26,7 +26,7 @@ stack_data randomWithCorrectingIterations(qa_global params) {
             //swap factories
             swap(&heuristics.assignment[i], &heuristics.assignment[j]);
 
-            double cost = assignmentCost(params, heuristics.assignment);
+            int cost = assignmentCost(params, heuristics.assignment);
 
             if (cost > heuristics.cost) {
                 swap(&heuristics.assignment[j], &heuristics.assignment[i]);
@@ -39,8 +39,8 @@ stack_data randomWithCorrectingIterations(qa_global params) {
     return heuristics;
 }
 
-double assignmentCost(qa_global params, const int *assignment) {
-    double cost = 0;
+int assignmentCost(qa_global params, const int *assignment) {
+    int cost = 0;
     for (int i = 0; i < params.locations; ++i) {
         for (int j = i; j < params.locations; ++j) {
             cost += params.distanceMatrix[i][j] * params.flowMatrix[assignment[i]][assignment[j]];
